docs(grid): depreaction story
